Collect artisan tests in a single class
